agent/requirements.txt
python-dotenv==1.1.0         
openai==1.73.0               
langchain==0.3.23           
langchain-openai==0.3.12    
langchain-community==0.3.21  
langchain-core==0.3.51
langchain-huggingface==0.1.2
langchain-text-splitters==0.3.8
langsmith==0.3.30            
sentence-transformers==4.0.2 
faiss-cpu==1.10.0           
tiktoken==0.9.0                  
scipy==1.13.1
tqdm==4.67.1                
beautifulsoup4==4.13.3
requests==2.32.3
pillow==11.2.1
transformers==4.51.2
torch==2.2.2
